backend/agents/requirements.txt
# Core LiveKit dependencies
livekit>=0.13.0
livekit-agents>=0.13.0

# AI Provider plugins
livekit-agents-plugin-openai>=0.8.0
livekit-plugins-silero>=0.8.0

# Additional utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
